# docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Event ingestion and analytics API with Redis deduplication, async worker batching, and ClickHouse metrics.",
        "title": "Insider One Case API",
        "contact": {},
        "version": "1.0.0"
    },
    "basePath": "/",
    "paths": {
        "/events": {
            "post": {
                "description": "Validates and enqueues one event for asynchronous processing.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "events"
                ],
                "summary": "Ingest a single event",
                "parameters": [
                    {
                        "description": "Event ingestion request",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.EventIngestRequest"
                        }
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Accepted",
                        "schema": {
                            "$ref": "#/definitions/model.EventIngestResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    },
                    "429": {
                        "description": "Too Many Requests",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/events/bulk": {
            "post": {
                "description": "Accepts between 1 and configured max events; returns aggregate outcome with partial-success semantics.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "events"
                ],
                "summary": "Ingest events in bulk",
                "parameters": [
                    {
                        "description": "Bulk event ingestion request",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.BulkEventIngestRequest"
                        }
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Accepted",
                        "schema": {
                            "$ref": "#/definitions/model.BulkEventIngestResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/health": {
            "get": {
                "description": "Returns basic service metadata and current UTC time.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "health"
                ],
                "summary": "Health check",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.HealthResponse"
                        }
                    }
                }
            }
        },
        "/metrics": {
            "get": {
                "description": "Returns totals and optional channel grouping for an event in the provided [from, to) unix timestamp range.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "metrics"
                ],
                "summary": "Query event metrics",
                "parameters": [
                    {
                        "type": "string",
                        "example": "purchase",
                        "description": "Event name",
                        "name": "event_name",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "example": 1710000000,
                        "description": "Inclusive unix timestamp lower bound",
                        "name": "from",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "example": 1710086400,
                        "description": "Exclusive unix timestamp upper bound",
                        "name": "to",
                        "in": "query",
                        "required": true
                    },
                    {
                        "enum": [
                            "channel"
                        ],
                        "type": "string",
                        "description": "Optional grouping",
                        "name": "group_by",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.MetricsResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "model.BulkEventIngestRequest": {
            "type": "object",
            "required": [
                "events"
            ],
            "properties": {
                "events": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.EventIngestRequest"
                    }
                }
            }
        },
        "model.BulkEventIngestResponse": {
            "type": "object",
            "properties": {
                "status": {
                    "description": "accepted_all, accepted_partial, rejected",
                    "type": "string"
                },
                "summary": {
                    "$ref": "#/definitions/model.BulkEventIngestSummary"
                }
            }
        },
        "model.BulkEventIngestSummary": {
            "type": "object",
            "properties": {
                "accepted": {
                    "type": "integer"
                },
                "duplicate": {
                    "type": "integer"
                },
                "error": {
                    "type": "integer"
                },
                "invalid": {
                    "type": "integer"
                },
                "overloaded": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
        "model.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "invalid_request"
                },
                "message": {
                    "type": "string",
                    "example": "malformed JSON body"
                }
            }
        },
        "model.EventIngestRequest": {
            "type": "object",
            "required": [
                "campaign_id",
                "channel",
                "event_name",
                "tags",
                "timestamp",
                "user_id"
            ],
            "properties": {
                "campaign_id": {
                    "type": "string",
                    "example": "cmp_123"
                },
                "channel": {
                    "type": "string",
                    "example": "mobile"
                },
                "event_name": {
                    "type": "string",
                    "example": "purchase"
                },
                "metadata": {
                    "type": "object"
                },
                "tags": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "promo",
                        "spring"
                    ]
                },
                "timestamp": {
                    "type": "integer",
                    "example": 1710000000
                },
                "user_id": {
                    "type": "string",
                    "example": "user_456"
                }
            }
        },
        "model.EventIngestResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "model.HealthData": {
            "type": "object",
            "properties": {
                "app": {
                    "type": "string",
                    "example": "insider-one-case"
                },
                "env": {
                    "type": "string",
                    "example": "development"
                },
                "status": {
                    "type": "string",
                    "example": "ok"
                },
                "time": {
                    "type": "string",
                    "example": "2026-03-19T22:50:00Z"
                }
            }
        },
        "model.HealthResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/model.HealthData"
                },
                "message": {
                    "type": "string",
                    "example": "ok"
                },
                "success": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
        "model.MetricsGroup": {
            "type": "object",
            "properties": {
                "count": {
                    "type": "integer"
                },
                "key": {
                    "type": "string"
                },
                "unique_users": {
                    "type": "integer"
                }
            }
        },
        "model.MetricsResponse": {
            "type": "object",
            "properties": {
                "event_name": {
                    "type": "string"
                },
                "from": {
                    "type": "integer"
                },
                "group_by": {
                    "type": "string"
                },
                "groups": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.MetricsGroup"
                    }
                },
                "to": {
                    "type": "integer"
                },
                "total_count": {
                    "type": "integer"
                },
                "unique_users": {
                    "type": "integer"
                }
            }
        }
    }
}